Refactor project layout. Include pytest.
