+++ BeansForDinner/Recipe/__init__.py
@@ -1,0 +1,4 @@
+from .recipe import Recipe
+from .recipe import recipe_title
+
+from .recipe_factory import RecipeFactory, AtomicRecipeFactory, CompositeRecipeFactory, CollectionFactory
