+++ BeansForDinner/Web/__init__.py
@@ -1,0 +1,37 @@
+import os
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
+from BeansForDinner.recipe import Recipe
+
+# init SQLAlchemy so we can use it later in our models
+db = SQLAlchemy()
+
+def create_app():
+    app = Flask(__name__)
+    app.config.from_pyfile('defaultconfig.py')
+    # app.config['DEBUG'] = True
+    # app.config['SECRET_KEY'] = 'updatethis'
+    # app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///../data/beans_db.sqlite'
+    db.init_app(app)
+
+    login_manager = LoginManager()
+    login_manager.login_view = 'auth.login'
+    login_manager.init_app(app)
+
+    from .models import User
+
+    @login_manager.user_loader
+    def load_user(user_id):
+        # since the user_id is just the primary key of our user table, use it in the query for the user
+        return User.query.get(int(user_id))
+    
+    # blueprint for auth routes in our app
+    from .auth import auth as auth_blueprint
+    app.register_blueprint(auth_blueprint)
+
+    # blueprint for non-auth parts of app
+    from .main import main as main_blueprint
+    app.register_blueprint(main_blueprint)
+
+    return app
