--- Recipe/__init__.py
@@ -1,3 +1,0 @@
-from .recipe import Recipe
-from .recipe import recipe_title
-from .recipe import SteelCutOats
